Swap the client/server connection to use WebSockets. - Refactor the WSS to its own file. - Remove unused, checked-in socket.io-client.js - Remove socket.io from package.json, as it's no longer needed.
