+++ package.json
@@ -32,7 +32,6 @@
     "random-js": "^2.1.0",
     "relaxed-json": "^1.0.0",
     "sharp": "^0.29.3",
-    "socket.io": "^2.3.0",
     "socket.io-client": "^2.3.0",
     "sortablejs": "^1.14.0",
     "three-full": "^11.3.2",
